Move dataset reading from preprocessing.py tp dataset.py
